+++ Source/Servers/ITCPServer.h
@@ -79,7 +79,7 @@ struct ITCPServer : public IServerEx
 
         return true;
     }
-    virtual bool onSocketstatus(int32_t *Readcount, size_t *Readsockets, int32_t *Writecount, size_t *Writesockets)
+    virtual bool onSocketstatus(size_t *Readcount, size_t *Readsockets, size_t *Writecount, size_t *Writesockets)
     {
         // We do not work with sockets in this version.
         return false;
